+++ lib/onix_flutter_mvvm.dart
@@ -2,8 +2,10 @@ library onix_flutter_mvvm;
 
 export 'src/command/command.dart';
 export 'src/action/action.dart';
-export 'src/extension/base_vm_builder_extension.dart';
-export 'src/extension/base_vm_consumer_extension.dart';
+export 'src/mixin/state_builder_mixin.dart';
+export 'src/mixin/state_consumer_mixin.dart';
+export 'src/mixin/widget_builder_mixin.dart';
+export 'src/mixin/widget_consumer_mixin.dart';
 export 'src/extension/vm_builder_extension.dart';
 export 'src/extension/vm_consumer_extension.dart';
 export 'src/view_model/view_model.dart' show ViewModel;